postgres_to_es/requirements.txt
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
distro==1.6.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.3
jsonschema==3.2.0
paramiko==2.8.0
psycopg2-binary==2.9.1
pycparser==2.21
pydantic==1.8.2
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dotenv==0.19.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
texttable==1.6.4
typing-extensions==3.10.0.2
urllib3==1.26.7
websocket-client==0.59.0

elasticsearch~=7.15.2